# public/chatgpt-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cultlike OS Executive Operating System API (ChatGPT & Claude Connectors)",
    "description": "Official API for ChatGPT Custom GPT Actions & Claude Connectors to orchestrate projects, sprint tasks, content vault deliverables, Fathom meetings, and execution analytics in Cultlike OS.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://callmy-mgmt.vercel.app",
      "description": "Cultlike OS Server"
    }
  ],
  "paths": {
...
    "/api/chatgpt/meetings": {
      "get": {
        "operationId": "listMeetings",
        "summary": "List and search Fathom video meetings",
        "description": "Fetches historical Fathom AI meeting recordings with filtering by keyword search, specific date (YYYY-MM-DD), attendee name, or limit.",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Search keyword matching title, attendee, or summary."
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter meetings on a specific date (YYYY-MM-DD)."
          },
          {
            "name": "attendee",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter meetings containing a specific attendee name."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 25
            },
            "description": "Max meetings to return."
          }
        ],
        "responses": {
          "200": {
            "description": "List of meetings."
          }
        }
      }
    },
...
    "/api/chatgpt/synthesize": {
      "post": {
        "operationId": "synthesizePlan",
        "summary": "Synthesize raw brain dump into phased project/task plan",
        "description": "Decomposes raw natural language brain dump or meeting notes into a phased project or task plan with optional auto-commit to database.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "text"
                ],
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "Brain dump text or meeting notes."
                  },
                  "auto_commit": {
                    "type": "boolean",
                    "default": false,
                    "description": "If true, automatically creates the project and tasks in the database."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan synthesized successfully."
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}